Fix CurrentMilliseconds overflow by using long
- Change CurrentMilliseconds from int to long in ITimeManager, FrozenTimeManager, EmptyTimeManager, and TimeManager
- Change FrozenStopwatch._startTime from int to long
- Log fatal and throw OverflowException if CurrentMilliseconds goes negative

+++ ControlBee/Interfaces/ITimeManager.cs
@@ -2,7 +2,7 @@
 
 public interface ITimeManager : IDisposable
 {
-    int CurrentMilliseconds { get; }
+    long CurrentMilliseconds { get; }
     void Sleep(int millisecondsTimeout);
     IStopwatch CreateWatch();
     void Register();
